interview agent restructuring
