--- Backend/interview_agent/.gitignore
@@ -1,27 +1,0 @@
-# ENV files (secrets)
-.env
-*.env
-backend/.env
-frontend/.env
-
-# Node
-node_modules/
-
-# Python
-__pycache__/
-*.pyc
-venv/
-.env/
-
-# Build folders
-dist/
-build/
-
-# Logs
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# OS junk
-.DS_Store
-Thumbs.db
