--- Backend/interview_agent/backend/app/main.py
@@ -1,35 +1,0 @@
-from fastapi import FastAPI
-from fastapi.middleware.cors import CORSMiddleware
-import uvicorn
-import os
-from dotenv import load_dotenv
-
-load_dotenv()
-
-app = FastAPI(title="RecruBotX API", version="0.1.0")
-
-# Configure CORS
-origins = ["*"]
-
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=origins,
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
-
-from app.api.websocket_manager import router as websocket_router
-
-app.include_router(websocket_router)
-
-@app.get("/")
-def read_root():
-    return {"message": "RecruBotX Interview Agent API is running"}
-
-@app.get("/health")
-def health_check():
-    return {"status": "healthy"}
-
-if __name__ == "__main__":
-    uvicorn.run("app.main:app", host="0.0.0.0", port=8000, reload=True)
