--- Backend/interview_agent/backend/reports/candidates/Ayyash_20260220_075757_report.json
@@ -1,29 +1,0 @@
-{
-    "score": 85,
-    "summary": "This is a placeholder report generated automatically as the interview concluded based on the 'finished' trigger. The candidate performed well.",
-    "skills": [
-        {
-            "name": "Communication",
-            "score": 90,
-            "feedback": "Excellent clarity."
-        },
-        {
-            "name": "Technical",
-            "score": 80,
-            "feedback": "Solid foundation."
-        },
-        {
-            "name": "Confidence",
-            "score": 85,
-            "feedback": "Very confident."
-        }
-    ],
-    "strengths": [
-        "Clear voice",
-        "Good pacing"
-    ],
-    "weaknesses": [
-        "Could elaborate more"
-    ],
-    "verdict": "Hire"
-}
