+++ Frontend/src/components/interview/DebugPanel.jsx
