+++ Frontend/src/components/interview/LiveInterviewSession.jsx
@@ -37,8 +37,7 @@ const LiveInterviewSession = ({
     candidateName,
     jobRole,
     onInterrupt,
-    liveCaption, /* Accept liveCaption from App parent */
-    onEndInterview
+    liveCaption /* Accept liveCaption from App parent */
 }) => {
     // State
     const [isTranscriptVisible, setIsTranscriptVisible] = useState(true);
@@ -184,7 +183,7 @@ const LiveInterviewSession = ({
 
                         <button
                             className="bg-red-500 hover:bg-red-600 text-white px-6 py-2.5 rounded-xl font-medium transition-colors text-sm"
-                            onClick={onEndInterview} // Mock end
+                            onClick={() => window.location.reload()} // Mock end
                         >
                             End Interview
                         </button>
